#13 GOT TRIANGLE RASTERIZATION WORKING FULLYgit add . had an error because of some int/float usage, as well as forgetting I changed the Vertex colour value to be a struct rather than the packed BGR values

+++ VertexCortex Purified/math.cpp
@@ -1,7 +1,7 @@
 #include "math.h"
 
 // 2d cross product (used in triangle rasterisation)
-int crossProduct(Vec2 AB, Vec2 AP) {
+float crossProduct(Vec2 AB, Vec2 AP) {
 
 	return AB.x * AP.y - AB.y * AP.x;
